requirements.txt
facenet-pytorch==2.5.3
Flask==2.3.2
gunicorn==21.2.0
joblib==1.3.1
numpy==1.25.1
opencv-python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
requests==2.31.0
scipy==1.11.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
Werkzeug==2.3.6